% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adviceAgrees}
\alias{adviceAgrees}
\title{Return whether the advice agrees with the initial decision}
\usage{
adviceAgrees(initial, advice)
}
\arguments{
\item{initial}{vector of initial decisions}

\item{advice}{vector of advisory estimates}
}
\description{
Return whether the advice agrees with the initial decision
}
