% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adviceCompatibility}
\alias{adviceCompatibility}
\title{Return a measure of how compatible advice is with an initial decision}
\usage{
adviceCompatibility(initial, advice)
}
\arguments{
\item{initial}{vector of initial decisions}

\item{advice}{vector of advisory estimates}
}
\value{
1 - error score
}
\description{
Return a measure of how compatible advice is with an initial decision
}
