% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adviseR.R
\docType{package}
\name{adviseR}
\alias{adviseR}
\title{adviseR: Advice Exchange Network Modelling}
\description{
Agent-based modelling of information exchange in a social network.
}
\section{Mypackage functions}{

The mypackage functions ...
}

