% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{advisorChoiceError}
\alias{advisorChoiceError}
\title{Calculate choice error averaged over the last 5 trials}
\usage{
advisorChoiceError(trust, advisorIndex, choice0, choice1, slope, nBack = 5L)
}
\arguments{
\item{trust}{Matrix of trust values - dims = trials x advisors}

\item{advisorIndex}{indices of chosen advisors}

\item{choice0}{indices of the advisor in the first choice slot}

\item{choice1}{indices of the advisor in the second choice slot}

\item{slope}{slope of the sigmoid function}

\item{nBack}{number of trials to look back. Looks at previous trials where
the advisorIndex advisor is in one of the choice slots.}
}
\value{
Matrix of mean pick rate for the advisorIndex advisor (column 1) and
mean predicted pick rate for that advisor (column 2)
}
\description{
Calculate choice error averaged over the last 5 trials
}
