% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{bayes}
\alias{bayes}
\title{Bayesian integration where the reliability of the advisor is the probability
the advisor agrees given we were correct.}
\usage{
bayes(initial, advice, weight, compression = 0.05)
}
\arguments{
\item{initial}{vector of initial decisions}

\item{advice}{vector of advisory estimates}

\item{weight}{trust rating for the advisor}

\item{compression}{whether to limit extreme values to c(x,1-x)}
}
\description{
Bayesian integration where the reliability of the advisor is the probability
the advisor agrees given we were correct.
}
\details{
Uses a Bayesian integration formula where
\deqn{c_2 = \frac{c_1 * t}{c_1 * t + (1-c_1)(1-t)}}{c2 = (c1*t)/(c1*t + (1-c1)(1-t))}
\eqn{c_2}{c2} is the final confidence (returned as a vector), and \eqn{c_1}{c1} the initial
confidence.
\eqn{t} is the probability of the advisor's advice given the initial decision
was correct. Where the advisor agrees, this is simply the trust we have in
the advisor (an advisor we trusted 100\% would always be expected to give the
same answer we did). Where the advisor disagrees, this is the opposite (we
consider it very unlikely a highly trusted advisor disagrees with us if we
are right).
}
