% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{biasGraph}
\alias{biasGraph}
\title{Correlation between bias and tie strength at each decision point}
\usage{
biasGraph(model)
}
\arguments{
\item{model}{network to calculate correlations for}
}
\value{
ggplot of correlations
}
\description{
Correlation between bias and tie strength at each decision point
}
