% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{biasToColourString}
\alias{biasToColourString}
\title{Return a colour string made up of r/g/b as desired with saturation scaled by
b}
\usage{
biasToColourString(
  b,
  colour = c("r", "g", "b"),
  maxVal = 2.5,
  minVal = 0,
  minSaturation = 0.1,
  maxSaturation = 0.9
)
}
\arguments{
\item{b}{vector of values to use for saturations}

\item{colour}{any of 'r', 'g', and 'b' to show channels used for saturation
(other channels are filled with FF)}

\item{maxVal}{maximum value for b}

\item{minVal}{minimum value for b}

\item{minSaturation}{minimum saturation value corresponding to minVal}

\item{maxSaturation}{maximum saturation value corresponding to maxVal}
}
\value{
character vector same length as b containing colour strings
}
\description{
Return a colour string made up of r/g/b as desired with saturation scaled by
b
}
