% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{detailGraphs}
\alias{detailGraphs}
\title{Add details to igraphs to hold information about connection activity and agent belief}
\usage{
detailGraphs(model)
}
\arguments{
\item{model}{network to apply details to}
}
\value{
model with updated $graphs
}
\description{
Add details to igraphs to hold information about connection activity and agent belief
}
