% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{.cluster_count}
\alias{.cluster_count}
\title{Return the number of clusters identified in a connectivity matrix by Ckmeans.1d.dp}
\usage{
.cluster_count(g, ..., .full = F)
}
\arguments{
\item{g}{connectivity igraph}

\item{...}{
  Arguments passed on to \code{\link[Ckmeans.1d.dp:Ckmeans.1d.dp]{Ckmeans.1d.dp::Ckmeans.1d.dp}}
  \describe{
    \item{\code{x}}{a numeric vector of data to be clustered. All \code{NA} elements must be removed from \code{x} before calling this function. The function will run faster on sorted \code{x} (in non-decreasing order) than an unsorted input.}
    \item{\code{k}}{either an exact integer number of clusters, or a vector of length two specifying the minimum and maximum numbers of clusters to be examined. The default is \code{c(1,9)}. When \code{k} is a range, the actual number of clusters is determined by Bayesian information criterion.}
    \item{\code{y}}{a value of 1 (default) to specify equal weights of 1 for each element in \code{x}, or a numeric vector of unequal non-negative weights for each element in \code{x}. It is highly recommended to use positive (instead of zero) weights to account for the influence of every element. The weights have a strong impact on the clustering result. When the number of clusters \code{k} is given as a range, the weights should be linearly scaled to sum up to the observed sample size. Currently, \code{Ckmedian.1d.dp} only works with an equal weight of 1.}
    \item{\code{method}}{a character string to specify the speedup method to the original cubic runtime dynamic programming. The default is \code{"linear"}. All methods generate the same optimal results but differ in runtime or memory usage. See Details.}
    \item{\code{estimate.k}}{a character string to specify the method to estimate optimal \code{k}. This argument is effective only when a range for \code{k} is provided. The default is \code{"BIC"}. See Details.}
  }}

\item{.full}{return the full Ckmeans.1d.dp output}
}
\description{
Return the number of clusters identified in a connectivity matrix by Ckmeans.1d.dp
}
