% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{.sensitivityCorrelation}
\alias{.sensitivityCorrelation}
\title{Calculate the strength of the sensitivity-advice weight correlation for in-
and out-degree}
\usage{
.sensitivityCorrelation(model)
}
\arguments{
\item{model}{network to calculate the correlations for}
}
\description{
Calculate the strength of the sensitivity-advice weight correlation for in-
and out-degree
}
