% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getConfidence}
\alias{getConfidence}
\title{Return the confidence of the \code{agents} given a percept, confidence
steepness, and bias.}
\usage{
getConfidence(agents, truth_sd)
}
\arguments{
\item{agents}{tbl with a snapshot of agents at a given decision time}

\item{truth_sd}{agents' beliefs about the world's variability}
}
\value{
agents' confidence the answer is 1 vs 0 as a proportion.
}
\description{
Return the confidence of the \code{agents} given a percept, confidence
steepness, and bias.
}
\details{
\code{agents} should have the fields \code{$percept},
\code{$confidence_slope}, and \code{$bias}
}
