% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getPercept}
\alias{getPercept}
\title{Return the percept of the \code{agents} given some truth}
\usage{
getPercept(agents)
}
\arguments{
\item{agents}{tbl with snapshot of agents at a given decision time}
}
\value{
agents' percepts
}
\description{
Return the percept of the \code{agents} given some truth
}
\details{
\code{agents} should have the fields \code{$truth} and
\code{$sensitivity}
}
