% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getUpdatedBias}
\alias{getUpdatedBias}
\title{Return vector of updated biases for agents}
\usage{
getUpdatedBias(agents, slope, limit = 0.05)
}
\arguments{
\item{agents}{tbl with a snapshot of agents at a given time}

\item{slope}{slope of the sigmoid function run on influence}

\item{limit}{clamp resulting bias to between limit and 1-limit}
}
\description{
Return vector of updated biases for agents
}
\details{
Bias is calculated by feeding influence (difference between final
and initial judgement) into a sigmoid function. This gives an amount of bias
influence between -.5 and .5, for which we take the absolute value and
subtract .5, giving a value between 0 and .5 which is around 0 for low levels
of influence and rises rapidly to around .5 for higher influence levels.
This bias influence is then scaled to between 0 and the agent's
bias_volatility parameter (when bias influence is .5), and this is used as
the amount the bias updates towards the final decision direction.
The final value of the bias is clamped to avoid extreme values.
}
