% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{graph_to_mat}
\alias{graph_to_mat}
\title{Return the matrix representation of the attributes of the edges of an igraph}
\usage{
graph_to_mat(g, attr = "weight")
}
\arguments{
\item{g}{connectivity igraph}

\item{attr}{attribute to use for values}
}
\description{
Return the matrix representation of the attributes of the edges of an igraph
}
