% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{groupRatio}
\alias{groupRatio}
\title{Return the ratio of weights of in-:out-group connections}
\usage{
groupRatio(g, original_groups = T)
}
\arguments{
\item{g}{graph for which to determine the ratio}

\item{original_groups}{whether to use the original groups (\code{T}) or
assign groups based on biases in the current graph (\code{F})}
}
\value{
double representing the ratio of mean weights in the graph for all
agents
}
\description{
Return the ratio of weights of in-:out-group connections
}
