% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{inspectModel}
\alias{inspectModel}
\title{Return a list with a variety of model graphs and stats}
\usage{
inspectModel(m)
}
\arguments{
\item{m}{model to inspect}
}
\description{
Return a list with a variety of model graphs and stats
}
