% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{newWeightsByDrift}
\alias{newWeightsByDrift}
\title{Return the new trust matrix for agents using weighted drift towards 0/1}
\usage{
newWeightsByDrift(agents, graph, confidence_weighted = T)
}
\arguments{
\item{agents}{tbl with a snapshot of agents at a given time}

\item{graph}{connectivity matrix of trust}

\item{confidence_weighted}{whether to use confidence to weight the updating or
instead rely only on dis/agreement}
}
\value{
Connectivity matrix of trust after accounting for agents' decisions
}
\description{
Return the new trust matrix for agents using weighted drift towards 0/1
}
\details{
Agents drift towards 0 or 1 trust according to whether or not the
advisor agreed with them. This can be weighted by confidence, too, where the
agreement matters more the more confident the agent was in their initial
decision. This latter is accomplished by a second drift from the original
trust rating to the prospective value.
}
