% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot an igraph of a network}
\usage{
plotGraph(model, i, activeColours = T)
}
\arguments{
\item{model}{network to plot}

\item{i}{generation to plot}

\item{activeColours}{whether to highlight the active advice connections}
}
\value{
plot object
}
\description{
Plot an igraph of a network
}
