% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run the simulation}
\usage{
runSimulation(
  n_agents = 6,
  n_decisions = 200,
  conf = T,
  bias_mean = 1,
  bias_sd = 1,
  sensitivity_sd = 1,
  trust_volatility_mean = 0.05,
  trust_volatility_sd = 0.01,
  bias_volatility_mean = 0.05,
  bias_volatility_sd = 0.01,
  starting_graph = NULL,
  randomSeed = NA,
  truth_fun = function(model, d) stats::rnorm(1),
  weighted_sampling = NA,
  asymptotic_confidence = c(0, 1)
)
}
\arguments{
\item{n_agents}{number of nodes in the network}

\item{n_decisions}{number of decisions to simulate}

\item{conf}{whether or not agents use confidence to update trust}

\item{bias_mean}{the mean for the agents' bias distribution (agents' biases
are drawn from normal distributions with mean +/- biasMean)}

\item{bias_sd}{standard deviation for the bias distribution}

\item{sensitivity_sd}{standard deviation for distribution of agents'
sensitivity (mean is 1)}

\item{trust_volatility_mean}{the mean volatility of agents' trust}

\item{trust_volatility_sd}{standard deviation}

\item{bias_volatility_mean}{the mean volatility of agents' biases (move this
proportion towards the final decision value from current bias at each step)}

\item{bias_volatility_sd}{standard deviation}

\item{starting_graph}{single number, vector, or n_agents-by-n_agents matrix
of starting trust weights between agents. Coerced to numeric}

\item{randomSeed}{the random seed to start the simulation with}

\item{truth_fun}{function taking the simulation and decision number as
arguments and returning the true state of the world as a single number}

\item{weighted_sampling}{a non-zero/NA value means agents choose who to seek
advice from according to how likely they are to trust the advice. The
weights are multiplied by this value (so values > 1 make source selection
more pronounced than advice weighting, and values < 1 make source selection
less pronounced than advice weighting). Negative values will make agents
actively seek out those they do not trust for advice.}

\item{asymptotic_confidence}{whether to constrain confidence in decision to
be between 0 (certain left) and 1 (certain right), as determined by a
sigmoid function. If this is a length 2 vector it contains the mean and SD
of the distribution from which agent sigmoid slopes are drawn. If this is
a function, it is called with the agents' details and should return the
slopes as a vector. \code{\link{abs}} is used to avoid negative slopes.
If this is FALSE, decision confidence is simply the agent's estimate of the
state of the world.}
}
\value{
a list with \itemize{
\item{"times"}{Timestamps associated with simulation stages.}
\item{"parameters"}{Parameters input into the model}
\item{"model"}{A list of model content \itemize{
\item{"agents"}{A tbl with 1 row for each decision made by an agent.}
\item{"graphs"}{Annotated igraphs of trust strength between agents at each
generation.}
}}
}
}
\description{
Run the simulation
}
