% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runSimulations}
\alias{runSimulations}
\title{Run a suite of simulations defined by params}
\usage{
runSimulations(params)
}
\arguments{
\item{params}{dataframe of parameters for simulations (see \code{runSimulation()} for details)}
}
\description{
Run a suite of simulations defined by params
}
