% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{selectAdvisorSimple}
\alias{selectAdvisorSimple}
\title{Return the advisor selections based on \code{graph}}
\usage{
selectAdvisorSimple(graph, weightedSelection = 0)
}
\arguments{
\item{graph}{weighted trust matrix for agents}

\item{weightedSelection}{scaling for probabilistic selection; 0 means
selection is equally weighted}
}
\value{
vector of the advisor id selected by each agent
}
\description{
Return the advisor selections based on \code{graph}
}
\details{
Weights the probability of selection of each advisor by using a
sigmoid with slope = weightedSelection on the trust in that advisor
}
