% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{sensitivityGraph}
\alias{sensitivityGraph}
\title{Correlation between sensitivity and tie strength}
\usage{
sensitivityGraph(model)
}
\arguments{
\item{model}{network to calculate correlations for}
}
\value{
ggplot of correlations
}
\description{
Correlation between sensitivity and tie strength
}
