% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{settingsStr}
\alias{settingsStr}
\title{A neat string of the parameters for a model for inclusion in graphs}
\usage{
settingsStr(model)
}
\arguments{
\item{model}{a simulation result model}
}
\description{
A neat string of the parameters for a model for inclusion in graphs
}
