% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmoid.inv}
\alias{sigmoid.inv}
\title{Inverse sigmoid function}
\usage{
sigmoid.inv(p, slope = 1)
}
\arguments{
\item{p}{probability}

\item{slope}{slope steepness parameter}
}
\description{
Inverse sigmoid function
}
