% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x, slope = 1, nudge = 1e-08)
}
\arguments{
\item{x}{value}

\item{slope}{slope steepness parameter}

\item{nudge}{adjustments applied if p is very close to 0 or 1}
}
\value{
probability of selecting positive over negative direction given x
}
\description{
Sigmoid function
}
