% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulationStep}
\alias{simulationStep}
\title{Each timestep agents are asked for a binary decision about whether a variable
is > 0. They form a noisy initial decision based on the true value plus their
bias plus noise. They then give their opinion to another agent, and receive
another agent's opinion as advice. The advice is integrated according to the
current weight placed on the other agent's trustworthiness. This weight is
then updated according to whether the agents agree, weighted by the
confidence of the deciding agent.}
\usage{
simulationStep(model, d)
}
\arguments{
\item{model}{to simulate the step for}

\item{d}{decision to simulate}
}
\value{
model updated to include values for decision d
}
\description{
Each timestep agents are asked for a binary decision about whether a variable
is > 0. They form a noisy initial decision based on the true value plus their
bias plus noise. They then give their opinion to another agent, and receive
another agent's opinion as advice. The advice is integrated according to the
current weight placed on the other agent's trustworthiness. This weight is
then updated according to whether the agents agree, weighted by the
confidence of the deciding agent.
}
