% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trustUpdate}
\alias{trustUpdate}
\title{Calculates the evolving trust scores based on agreement}
\usage{
trustUpdate(trust, advisorId, advisorAgrees, updateRate)
}
\arguments{
\item{trust}{Matrix of trust values - dims = trials x advisors}

\item{advisorId}{ids of the advisor giving advice on each trial}

\item{advisorAgrees}{whether advisor agrees on each trial}

\item{updateRate}{amount to down/upweight trust given dis/agreement}
}
\value{
New trust values after iterating through and up/downweighting trust
}
\description{
Calculates the evolving trust scores based on agreement
}
