% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{weighted}
\alias{weighted}
\title{Weighted average where \code{a} uses \code{weight_a} and \code{b} uses
\code{1 - weight_a}}
\usage{
weighted(a, b, weight_a)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{weight_a}{weight given to a (1-\code{weight_a}) is used for \code{b}}
}
\description{
Weighted average where \code{a} uses \code{weight_a} and \code{b} uses
\code{1 - weight_a}
}
