% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{weightEvolution}
\alias{weightEvolution}
\title{Plot evolving histograms for the advice weights of a model}
\usage{
weightEvolution(
  model,
  decisions = function(d) as.integer(seq(1, length(d), length.out = 10)),
  binwidth = 0.025,
  k = 1:2
)
}
\arguments{
\item{model}{model to inspect}

\item{decisions}{function taking an integer vector of decision numbers and
returning a logical vector of whether to plot weights at that decision}

\item{binwidth}{width of the histogram bins}

\item{k}{cluster counts to compare with .cluster_count}
}
\description{
Plot evolving histograms for the advice weights of a model
}
