% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{weightToColourString}
\alias{weightToColourString}
\title{Calculate the colour strings for different weights}
\usage{
weightToColourString(graph)
}
\arguments{
\item{graph}{with weights to use}
}
\value{
list of #RRGGBB codes with colours for the weights
}
\description{
Calculate the colour strings for different weights
}
