% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{biasGraph}
\alias{biasGraph}
\title{Correlation between bias and tie strength at each decision point}
\usage{
biasGraph(model, use_starting_bias = F)
}
\arguments{
\item{model}{network to calculate correlations for}

\item{use_starting_bias}{whether to use the starting bias as opposed to the
current bias at each timestep}
}
\value{
ggplot of correlations
}
\description{
Correlation between bias and tie strength at each decision point
}
