% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{.biasCorrelation}
\alias{.biasCorrelation}
\title{Calculation of bias-tie strength correlations}
\usage{
.biasCorrelation(model)
}
\arguments{
\item{model}{network whose correlations should be calculated}
}
\description{
Calculation of bias-tie strength correlations
}
