% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-agents.R
\name{makeAgents}
\alias{makeAgents}
\title{Agents are constructed with a bias on their binary decision (constant over
time) and a set of weights governing how seriously the advice of other agents
is taken (modified over time).}
\usage{
makeAgents(
  n = list(p = 6, d = 200),
  biasMean = 1,
  biasSD = 1,
  sensitivitySD = 1
)
}
\arguments{
\item{n}{list of numbers (participants and decisions) to simulate}

\item{biasMean}{the mean for the agents' bias distribution (agents' biases
are drawn from normal distributions with mean +/- biasMean)}

\item{biasSD}{standard deviation for the bias distribution}

\item{sensitivitySD}{standard deviation for distribution of agents'
sensitivity (mean is 1)}
}
\value{
\code{list(
  agents = tibble of agents' decisions, advice, etc. at each time point
  graphs = list of agents' trust matrix for each time point
)}
}
\description{
Agents are constructed with a bias on their binary decision (constant over
time) and a set of weights governing how seriously the advice of other agents
is taken (modified over time).
}
