% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-agents.R
\name{makeAgents}
\alias{makeAgents}
\title{Agents are constructed with a bias on their binary decision (constant over
time) and a set of weights governing how seriously the advice of other agents
is taken (modified over time).}
\usage{
makeAgents(
  n_agents = n_agents,
  n_decisions = n_decisions,
  bias_mean = 0,
  bias_sd = 1,
  sensitivity_sd = 1,
  trust_volatility_mean = 0.05,
  trust_volatility_sd = 0.01,
  bias_volatility_mean = 0.05,
  bias_volatility_sd = 0.01,
  confidence_slope_mean = 1,
  confidence_slope_sd = 0,
  weighted_sampling_mean = 0,
  weighted_sampling_sd = 0,
  starting_graph = NULL
)
}
\arguments{
\item{n_agents}{number of agents to create}

\item{n_decisions}{number of decisions}

\item{bias_mean}{the mean for the agents' bias distribution (agents' biases
are drawn from normal distributions with mean +/- biasMean). Translated via
sigmoid function to between 0 and 1, and represents the prior probability
that the answer is 1.}

\item{bias_sd}{standard deviation for the bias distribution}

\item{sensitivity_sd}{standard deviation for distribution of agents'
sensitivity (mean is 1)}

\item{trust_volatility_mean}{the mean volatility of agents' trust}

\item{trust_volatility_sd}{standard deviation}

\item{bias_volatility_mean}{the mean volatility of agents' biases (move this
proportion towards the final decision value from current bias at each step)}

\item{bias_volatility_sd}{standard deviation}

\item{confidence_slope_mean}{the mean of the distribution from which agents
take their slopes for the sigmoid function mapping continuous evidence to
a probability of a categorical decision.}

\item{confidence_slope_sd}{standard deviation}

\item{weighted_sampling_mean}{a non-zero value means agents choose who to
seek advice from according to how likely they are to trust the advice. The
weights are raised to the power of this value (so values > 1 make source
selection more pronounced than advice weighting, and values < 1 make source
selection less pronounced than advice weighting). Negative values will make
agents actively seek out those they do not trust for advice.}

\item{weighted_sampling_sd}{standard deviation}

\item{starting_graph}{single number, vector, or n_agents-by-n_agents matrix
of starting trust weights between agents. Coerced to numeric. Can also be
a function taking the first generation of the agents tbl as an input and
returning an n-by-n matrix of trust values between 0 and 1, where n is the
number of agents, 0 represents completely untrustworthy, .5 random, and 1
completely trustworthy.}
}
\value{
\code{list(
  agents = tibble of agents' decisions, advice, etc. at each time point
  graphs = list of agents' trust matrix for each time point
)}
}
\description{
Agents are constructed with a bias on their binary decision (constant over
time) and a set of weights governing how seriously the advice of other agents
is taken (modified over time).
}
\details{
the \code{agents} tibble is an n_agents*n_decisions by 12 table with
\itemize{
\item{"id"}{The agent's identifier}
\item{"decision"}{The decision number}
\item{"sensitivity"}{The agent's ability to do the task}
\item{"trust_volatility"}{How quickly the agent's trust updates}
\item{"bias_volatility"}{How quickly the agent's bias updates}
\item{"weighted_sampling"}{How heavily trust governs advice sampling behaviour}
\item{"bias"}{The agent's (initial) bias}
\item{"truth"}{The true state of the world (same for a given decision for all agents)}
\item{"initial"}{The agent's initial estimate of the truth}
\item{"advisor"}{ID of the agent whose advice is received}
\item{"weight"}{The weight assigned to the advice by the agent}
\item{"final"}{The final decision of the agent}
}
}
