% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{networkGraph}
\alias{networkGraph}
\title{Plot multiple graphs of network connectivity to show network evolution}
\usage{
networkGraph(model, timepoints = 2, ...)
}
\arguments{
\item{model}{network}

\item{timepoints}{either an integer representing the number of evenly-spaced
time points (including first and final), or an integer vector giving the
time points to inspect.}

\item{...}{
  Arguments passed on to \code{\link[=plotGraph]{plotGraph}}
  \describe{
    \item{\code{i}}{generation to plot}
    \item{\code{activeColours}}{whether to highlight the active advice connections}
    \item{\code{layout}}{layout for the graph}
  }}
}
\value{
NULL (invisible)
}
\description{
Plot multiple graphs of network connectivity to show network evolution
}
