% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{newWeightsBayes}
\alias{newWeightsBayes}
\title{Return new trust matrix updating using Bayes rule}
\usage{
newWeightsBayes(agents, graph, confidence_weighted, compression = 0.05)
}
\description{
Return new trust matrix updating using Bayes rule
}
