% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot an igraph of a network}
\usage{
plotGraph(model, i, activeColours = T, layout = igraph::layout_nicely, ...)
}
\arguments{
\item{model}{network to plot}

\item{i}{generation to plot}

\item{activeColours}{whether to highlight the active advice connections}

\item{layout}{layout for the graph}

\item{...}{
  Arguments passed on to \code{\link[igraph:plot.igraph]{igraph::plot.igraph}}
  \describe{
    \item{\code{x}}{The graph to plot.}
    \item{\code{axes}}{Logical, whether to plot axes, defaults to FALSE.}
    \item{\code{add}}{Logical scalar, whether to add the plot to the current device, or
delete the device's current contents first.}
    \item{\code{xlim}}{The limits for the horizontal axis, it is unlikely that you want
to modify this.}
    \item{\code{ylim}}{The limits for the vertical axis, it is unlikely that you want
to modify this.}
    \item{\code{mark.groups}}{A list of vertex id vectors. It is interpreted as a set
of vertex groups. Each vertex group is highlighted, by plotting a colored
smoothed polygon around and \dQuote{under} it. See the arguments below to
control the look of the polygons.}
    \item{\code{mark.shape}}{A numeric scalar or vector. Controls the smoothness of the
vertex group marking polygons. This is basically the \sQuote{shape}
parameter of the \code{\link[graphics]{xspline}} function, its possible
values are between -1 and 1. If it is a vector, then a different value is
used for the different vertex groups.}
    \item{\code{mark.col}}{A scalar or vector giving the colors of marking the
polygons, in any format accepted by \code{\link[graphics]{xspline}}; e.g.
numeric color ids, symbolic color names, or colors in RGB.}
    \item{\code{mark.border}}{A scalar or vector giving the colors of the borders of
the vertex group marking polygons. If it is \code{NA}, then no border is
drawn.}
    \item{\code{mark.expand}}{A numeric scalar or vector, the size of the border around
the marked vertex groups. It is in the same units as the vertex sizes. If a
vector is given, then different values are used for the different vertex
groups.}
  }}
}
\value{
plot object
}
\description{
Plot an igraph of a network
}
