% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run the simulation}
\usage{
runSimulation(
  n_agents = 6,
  n_decisions = 200,
  conf = T,
  biasMean = 1,
  biasSD = 1,
  sensitivitySD = 1,
  learningRate = 0.1,
  randomSeed = NA
)
}
\arguments{
\item{n_agents}{number of nodes in the network}

\item{n_decisions}{number of decisions to simulate}

\item{conf}{whether or not agents use confidence to update trust}

\item{biasMean}{the mean for the agents' bias distribution (agents' biases
are drawn from normal distributions with mean +/- biasMean)}

\item{biasSD}{standard deviation for the bias distribution}

\item{sensitivitySD}{standard deviation for distribution of agents'
sensitivity (mean is 1)}

\item{learningRate}{size of the trust update at each time step}

\item{randomSeed}{the random seed to start the simulation with}
}
\value{
a list with \itemize{
\item{"times"}{Timestamps associated with simulation stages.}
\item{"parameters"}{Parameters input into the model}
\item{"model"}{A list of model content \itemize{
\item{"agents"}{A tbl with 1 row for each decision made by an agent.}
\item{"graphs"}{Annotated igraphs of trust strength between agents at each
generation.}
}}
}
}
\description{
Run the simulation
}
