% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x, slope = 1)
}
\arguments{
\item{x}{value}

\item{slope}{slope steepness parameter}
}
\value{
probability of selecting positive over negative direction given x
}
\description{
Sigmoid function
}
