% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_from_data.R
\name{simulateFromData}
\alias{simulateFromData}
\title{Simulate interactions based on specified data and parameters}
\usage{
simulateFromData(d, params, detailed_output = F)
}
\arguments{
\item{d}{tbl of data with columns 'advisorIndex', 'choice0', 'choice1',
'advisorAgrees', 'confidenceShift'}

\item{params}{1x2 tbl of parameters 'confidence slope' and 'trust update
rate'}

\item{detailed_output}{whether to return an augmented \code{d} tbl (\code{T})
or a named numeric vector of Mean Squared Errors}
}
\value{
Mean Squared Error of simulation with parameters compared to
empirical data values or augmented \code{d} argument, depending on whether
\code{detailed_output} is requested
}
\description{
Simulate interactions based on specified data and parameters
}
